fix security delete

+++ backend/controllers/stuff.js
@@ -25,9 +25,24 @@ exports.modifySauce = (req, res, next) => {
 }
 
 exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({_id: req.params.id})
-        .then(() => res.status(200).json({ message: 'Sauce supprimé !'}))
-        .catch(error => res.status(400).json({ error }))
+    Sauce.findOne({_id: req.params.id})
+      .then(
+        (sauce) => {
+          if (!sauce) {
+            return res.status(404).json({
+              error: new Error ('Objet non trouvé !')
+            });
+          }
+          if (sauce.userId !== req.auth.userId) {
+            return res.status(400).json({
+              error: new Error('Requete non autorisée !')
+            })
+          }
+          Sauce.deleteOne({_id: req.params.id})
+            .then(() => res.status(200).json({ message: 'Sauce supprimé !'}))
+            .catch(error => res.status(400).json({ error }))
+        }
+      ) 
 }
 
 exports.getOneSauce = (req, res, next) => {
